+++ vm/src/types/structseq.rs
@@ -46,24 +46,25 @@ pub trait PyStructSequence: StaticType + PyClassImpl + Sized + 'static {
             let s = value.repr(vm)?;
             Ok(format!("{}={}", name, s))
         };
-        let (body, suffix) =
-            if let Some(_guard) = rustpython_vm::vm::ReprGuard::enter(vm, zelf.as_object()) {
-                if Self::FIELD_NAMES.len() == 1 {
-                    let value = zelf.as_slice().first().unwrap();
-                    let formatted = format_field((value, Self::FIELD_NAMES[0]))?;
-                    (formatted, ",")
-                } else {
-                    let fields: PyResult<Vec<_>> = zelf
-                        .as_slice()
-                        .iter()
-                        .zip(Self::FIELD_NAMES.iter().copied())
-                        .map(format_field)
-                        .collect();
-                    (fields?.join(", "), "")
-                }
+        let (body, suffix) = if let Some(_guard) =
+            rustpython_vm::recursion::ReprGuard::enter(vm, zelf.as_object())
+        {
+            if Self::FIELD_NAMES.len() == 1 {
+                let value = zelf.as_slice().first().unwrap();
+                let formatted = format_field((value, Self::FIELD_NAMES[0]))?;
+                (formatted, ",")
             } else {
-                (String::new(), "...")
-            };
+                let fields: PyResult<Vec<_>> = zelf
+                    .as_slice()
+                    .iter()
+                    .zip(Self::FIELD_NAMES.iter().copied())
+                    .map(format_field)
+                    .collect();
+                (fields?.join(", "), "")
+            }
+        } else {
+            (String::new(), "...")
+        };
         Ok(format!("{}({}{})", Self::TP_NAME, body, suffix))
     }
